+++ src/auth/auth.controller.ts
@@ -1,10 +1,10 @@
 import { Controller, Get, Query, Res } from '@nestjs/common';
 import { Response } from 'express';
-import { AuthService } from './auth.service';
+import { KakaoService } from './kakao.service';
 
 @Controller('auth')
 export class AuthController {
-  constructor(private readonly authService: AuthService) {}
+  constructor(private readonly authService: KakaoService) {}
 
   @Get('kakao')
   kakaoLogin(@Res() res: Response) {
